Add SponsoredDataset as an abstruction.
