+++ generator/booking_dataset.cpp
@@ -3,133 +3,46 @@
 #include "generator/booking_scoring.hpp"
 #include "generator/feature_builder.hpp"
 
-#include "platform/local_country_file_utils.hpp"
-#include "platform/platform.hpp"
-
 #include "indexer/classificator.hpp"
 #include "indexer/ftypes_matcher.hpp"
 
-#include "geometry/distance_on_sphere.hpp"
-
-#include "base/logging.hpp"
 #include "base/string_utils.hpp"
-
-#include "std/fstream.hpp"
-#include "std/iostream.hpp"
-#include "std/sstream.hpp"
 
 namespace generator
 {
-namespace
-{
-string EscapeTabs(string const & str)
-{
-  stringstream ss;
-  for (char c : str)
-  {
-    if (c == '\t')
-      ss << "\\t";
-    else
-      ss << c;
-  }
-  return ss.str();
-}
-}  // namespace
+// BookingDataset::AddressMatcher::AddressMatcher()
+// {
+//   vector<platform::LocalCountryFile> localFiles;
 
-BookingDataset::BookingId const BookingDataset::kInvalidHotelIndex = BookingId(numeric_limits<BookingId::RepType>::max());
+//   Platform & platform = GetPlatform();
+//   platform::FindAllLocalMapsInDirectoryAndCleanup(platform.WritableDir(), 0 /* version */,
+//                                                   -1 /* latestVersion */, localFiles);
 
-BookingDataset::Hotel::Hotel(string const & src)
-{
-  vector<string> rec;
-  strings::ParseCSVRow(src, '\t', rec);
-  CHECK(rec.size() == FieldsCount(), ("Error parsing hotels.tsv line:", EscapeTabs(src)));
+//   for (platform::LocalCountryFile const & localFile : localFiles)
+//   {
+//     LOG(LINFO, ("Found mwm:", localFile));
+//     try
+//     {
+//       m_index.RegisterMap(localFile);
+//     }
+//     catch (RootException const & ex)
+//     {
+//       CHECK(false, ("Bad mwm file:", localFile));
+//     }
+//   }
 
-  strings::to_uint(rec[Index(Fields::Id)], id.Get());
-  strings::to_double(rec[Index(Fields::Latitude)], lat);
-  strings::to_double(rec[Index(Fields::Longtitude)], lon);
+//   m_coder = make_unique<search::ReverseGeocoder>(m_index);
+// }
 
-  name = rec[Index(Fields::Name)];
-  address = rec[Index(Fields::Address)];
+// void BookingDataset::AddressMatcher::operator()(Hotel & hotel)
+// {
+//   search::ReverseGeocoder::Address addr;
+//   m_coder->GetNearbyAddress(MercatorBounds::FromLatLon(hotel.lat, hotel.lon), addr);
+//   hotel.street = addr.GetStreetName();
+//   hotel.houseNumber = addr.GetHouseNumber();
+// }
 
-  strings::to_uint(rec[Index(Fields::Stars)], stars);
-  strings::to_uint(rec[Index(Fields::PriceCategory)], priceCategory);
-  strings::to_double(rec[Index(Fields::RatingBooking)], ratingBooking);
-  strings::to_double(rec[Index(Fields::RatingUsers)], ratingUser);
-
-  descUrl = rec[Index(Fields::DescUrl)];
-
-  strings::to_uint(rec[Index(Fields::Type)], type);
-
-  translations = rec[Index(Fields::Translations)];
-}
-
-ostream & operator<<(ostream & s, BookingDataset::Hotel const & h)
-{
-  s << fixed << setprecision(7);
-  return s << "Id: " << h.id << "\t Name: " << h.name << "\t Address: " << h.address
-           << "\t lat: " << h.lat << " lon: " << h.lon;
-}
-
-BookingDataset::AddressMatcher::AddressMatcher()
-{
-  vector<platform::LocalCountryFile> localFiles;
-
-  Platform & platform = GetPlatform();
-  platform::FindAllLocalMapsInDirectoryAndCleanup(platform.WritableDir(), 0 /* version */,
-                                                  -1 /* latestVersion */, localFiles);
-
-  for (platform::LocalCountryFile const & localFile : localFiles)
-  {
-    LOG(LINFO, ("Found mwm:", localFile));
-    try
-    {
-      m_index.RegisterMap(localFile);
-    }
-    catch (RootException const & ex)
-    {
-      CHECK(false, ("Bad mwm file:", localFile));
-    }
-  }
-
-  m_coder = make_unique<search::ReverseGeocoder>(m_index);
-}
-
-void BookingDataset::AddressMatcher::operator()(Hotel & hotel)
-{
-  search::ReverseGeocoder::Address addr;
-  m_coder->GetNearbyAddress(MercatorBounds::FromLatLon(hotel.lat, hotel.lon), addr);
-  hotel.street = addr.GetStreetName();
-  hotel.houseNumber = addr.GetHouseNumber();
-}
-
-BookingDataset::BookingDataset(string const & dataPath, string const & addressReferencePath)
-{
-  if (dataPath.empty())
-    return;
-
-  ifstream dataSource(dataPath);
-  if (!dataSource.is_open())
-  {
-    LOG(LERROR, ("Error while opening", dataPath, ":", strerror(errno)));
-    return;
-  }
-
-  LoadHotels(dataSource, addressReferencePath);
-}
-
-BookingDataset::BookingDataset(istream & dataSource, string const & addressReferencePath)
-{
-  LoadHotels(dataSource, addressReferencePath);
-}
-
-BookingDataset::BookingId BookingDataset::GetMatchingHotelId(FeatureBuilder1 const & fb) const
-{
-  if (CanBeBooking(fb))
-    return MatchWithBooking(fb);
-  return kInvalidHotelIndex;
-}
-
-bool BookingDataset::CanBeBooking(FeatureBuilder1 const & fb) const
+bool BookingDataset::NecessaryMatchingConditionHolds(FeatureBuilder1 const & fb) const
 {
   if (fb.GetName(StringUtf8Multilang::kDefaultCode).empty())
     return false;
@@ -137,49 +50,8 @@ bool BookingDataset::CanBeBooking(FeatureBuilder1 const & fb) const
   return ftypes::IsHotelChecker::Instance()(fb.GetTypes());
 }
 
-BookingDataset::Hotel const & BookingDataset::GetHotelById(BookingId const id) const
-{
-  auto const it = m_hotels.find(id);
-  CHECK(it != end(m_hotels), ("Got wrong hotel id:", id));
-  return it->second;
-}
-
-BookingDataset::Hotel & BookingDataset::GetHotelById(BookingId const id)
-{
-  auto it = m_hotels.find(id);
-  CHECK(it != end(m_hotels), ("Got wrong hotel id:", id));
-  return it->second;
-}
-
-vector<BookingDataset::BookingId> BookingDataset::GetNearestHotels(
-    ms::LatLon const & latLon, size_t const limit,
-    double const maxDistance /* = 0.0 */) const
-{
-  namespace bgi = boost::geometry::index;
-
-  vector<BookingId> indexes;
-  for_each(bgi::qbegin(m_rtree, bgi::nearest(TPoint(latLon.lat, latLon.lon), limit)),
-           bgi::qend(m_rtree), [this, &latLon, &indexes, maxDistance](TValue const & v)
-           {
-             auto const & hotel = GetHotelById(v.second);
-             double const dist = ms::DistanceOnEarth(latLon.lat, latLon.lon, hotel.lat, hotel.lon);
-             if (maxDistance != 0.0 && dist > maxDistance /* max distance in meters */)
-               return;
-
-             indexes.emplace_back(v.second);
-           });
-
-  return indexes;
-}
-
-void BookingDataset::BuildHotels(function<void(FeatureBuilder1 &)> const & fn) const
-{
-  for (auto const & item : m_hotels)
-    BuildHotel(item.second, fn);
-}
-
-void BookingDataset::BuildHotel(Hotel const & hotel,
-                                function<void(FeatureBuilder1 &)> const & fn) const
+void BookingDataset::BuildObject(SponsoredDataset::Object const & hotel,
+                                 function<void(FeatureBuilder1 &)> const & fn) const
 {
   FeatureBuilder1 fb;
   FeatureParams params;
@@ -276,68 +148,23 @@ void BookingDataset::BuildHotel(Hotel const & hotel,
   fn(fb);
 }
 
-void BookingDataset::LoadHotels(istream & src, string const & addressReferencePath)
-{
-  m_hotels.clear();
-  m_rtree.clear();
-
-  for (string line; getline(src, line);)
-  {
-    Hotel hotel(line);
-    m_hotels.emplace(hotel.id, hotel);
-  }
-
-  if (!addressReferencePath.empty())
-  {
-    LOG(LINFO, ("Reference addresses for booking objects", addressReferencePath));
-    Platform & platform = GetPlatform();
-    string const backupPath = platform.WritableDir();
-    platform.SetWritableDirForTests(addressReferencePath);
-
-    AddressMatcher addressMatcher;
-
-    size_t matchedNum = 0;
-    size_t emptyAddr = 0;
-    for (auto & item : m_hotels)
-    {
-      auto & hotel = item.second;
-      addressMatcher(hotel);
-
-      if (hotel.address.empty())
-        ++emptyAddr;
-      if (hotel.IsAddressPartsFilled())
-        ++matchedNum;
-    }
-    LOG(LINFO,
-        ("Num of hotels:", m_hotels.size(), "matched:", matchedNum, "empty addresses:", emptyAddr));
-    platform.SetWritableDirForTests(backupPath);
-  }
-
-  for (auto const & item : m_hotels)
-  {
-    auto const & hotel = item.second;
-    TBox b(TPoint(hotel.lat, hotel.lon), TPoint(hotel.lat, hotel.lon));
-    m_rtree.insert(make_pair(b, hotel.id));
-  }
-}
-
-BookingDataset::BookingId BookingDataset::MatchWithBooking(FeatureBuilder1 const & fb) const
+BookingDataset::ObjectId BookingDataset::FindMatchingObjectIdImpl(FeatureBuilder1 const & fb) const
 {
   auto const name = fb.GetName(StringUtf8Multilang::kDefaultCode);
 
   if (name.empty())
-    return kInvalidHotelIndex;
+    return kInvalidObjectId;
 
   // Find |kMaxSelectedElements| nearest values to a point.
-  auto const bookingIndexes = GetNearestHotels(MercatorBounds::ToLatLon(fb.GetKeyPoint()),
-                                               kMaxSelectedElements, kDistanceLimitInMeters);
+  auto const bookingIndexes = GetNearestObjects(MercatorBounds::ToLatLon(fb.GetKeyPoint()),
+                                                kMaxSelectedElements, kDistanceLimitInMeters);
 
   for (auto const j : bookingIndexes)
   {
-    if (booking_scoring::Match(GetHotelById(j), fb).IsMatched())
+    if (booking_scoring::Match(GetObjectById(j), fb).IsMatched())
       return j;
   }
 
-  return kInvalidHotelIndex;
+  return kInvalidObjectId;
 }
 }  // namespace generator
