+++ generator/generator.pro
@@ -36,6 +36,7 @@ SOURCES += \
     region_meta.cpp \
     routing_generator.cpp \
     search_index_builder.cpp \
+    sponsored_dataset.cpp \
     srtm_parser.cpp \
     statistics.cpp \
     tesselator.cpp \
@@ -72,6 +73,7 @@ HEADERS += \
     region_meta.hpp \
     routing_generator.hpp \
     search_index_builder.hpp \
+    sponsored_dataset.hpp \
     srtm_parser.hpp \
     statistics.hpp \
     tag_admixer.hpp \
